Move hash.hpp into own folder
