+++ BabelWiresLib/Features/recordFeature.cpp
@@ -10,7 +10,7 @@
 #include <BabelWiresLib/Features/modelExceptions.hpp>
 
 #include <Common/Identifiers/identifierRegistry.hpp>
-#include <Common/Utilities/hash.hpp>
+#include <Common/Hash/hash.hpp>
 #include <Common/types.hpp>
 
 int babelwires::RecordFeature::getNumFeatures() const {
